server/python/requirements.txt
async-timeout==2.0.1
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.4
gunicorn==19.7.1
idna==2.6
idna-ssl==1.1.0
itsdangerous==0.24
Jinja2==2.10
livereload==2.5.1
MarkupSafe==1.0
more-itertools==4.1.0
multidict==4.1.0
pluggy==0.6.0
py==1.5.3
pytest==3.5.0
python-dotenv==0.8.2
requests==2.20
six==1.11.0
stripe==2.12.0
tornado==5.0.2
urllib3==1.22
Werkzeug==0.14.1
yarl==1.1.1
